Refactor: Progress Bar 컴포넌트 분리
