+++ src/components/Main.jsx
@@ -3,6 +3,7 @@ import { faPaw, faCirclePlus } from "@fortawesome/free-solid-svg-icons";
 import styled from "styled-components";
 import { useEffect, useState } from "react";
 import TodoItem from "./TodoItem";
+import Progress from "./Progress";
 
 const Main = () => {
   const initialData = localStorage.getItem("data")
@@ -32,9 +33,6 @@ const Main = () => {
     setValue("");
   };
 
-  const doneCount = data.filter((el) => el.isDone).length;
-  const totalCount = data.length;
-
   const toggleTodo = (id) => {
     const updatedData = data.map((el) =>
       el.id === id ? { ...el, isDone: !el.isDone } : el
@@ -63,20 +61,7 @@ const Main = () => {
         </SubmitBtn>
       </InputWrapper>
 
-      <ProgressWrapper>
-        <ProgressText>
-          {totalCount !== 0 && doneCount === totalCount
-            ? "최고예요"
-            : "달려봐요"}{" "}
-          !
-        </ProgressText>
-        <ProgressBar>
-          <ProgressBarDone
-            $ratio={(doneCount / totalCount) * 100}
-          ></ProgressBarDone>
-        </ProgressBar>
-        <ProgressCount>{doneCount + " / " + totalCount}</ProgressCount>
-      </ProgressWrapper>
+      <Progress data={data} />
 
       <TodoWrapper>
         <div>
@@ -144,36 +129,6 @@ const Input = styled.input`
 `;
 
 const SubmitBtn = styled.button``;
-
-// Prgress Bar
-const ProgressWrapper = styled.section`
-  display: grid;
-  grid-template-columns: auto 1fr auto;
-  align-items: center;
-  column-gap: 12px;
-  padding: 0px 10px;
-  font-size: 22px;
-`;
-
-const ProgressText = styled.span``;
-
-const ProgressBar = styled.span`
-  background-color: #c0c0c0;
-  border-radius: 10px;
-  height: 20px;
-  position: relative;
-`;
-
-const ProgressBarDone = styled.span`
-  background-color: #7f7f7f;
-  border-radius: 10px;
-  height: inherit;
-  position: absolute;
-  transition: width 0.4s ease-in-out;
-  width: ${(props) => props.$ratio}%;
-`;
-
-const ProgressCount = styled.span``;
 
 //Todo
 const TodoWrapper = styled.section`
